refactor: :recycle: adding enum in payment
