+++ src/enums/enumPaymentType.ts
@@ -1,0 +1,4 @@
+export enum PaymentType {
+ PIX = 'PIX',
+ BILLET = 'BILLET',
+}
