+++ src/models/modelPaymentBillet.ts
@@ -1,7 +1,8 @@
-import { PaymentType } from "src/interfaces/Payment";
+import { Payment } from "src/interfaces/Payment";
+import { PaymentType } from "src/enums/enumPaymentType";
 
-export class PaymentBillet implements PaymentType {
- type: 'BILLET'
+export class PaymentBillet implements Payment {
+ type: PaymentType.BILLET
 
  pay(amountPayment: number, paymentReference: string ): void {}
