+++ src/models/modelPaymentPIX.ts
@@ -1,7 +1,8 @@
-import { PaymentType } from "src/interfaces/Payment";
+import { Payment } from "src/interfaces/Payment";
+import { PaymentType } from "src/enums/enumPaymentType";
 
-export class PaymentPix implements PaymentType {
- type: 'PIX'
+export class PaymentPix implements Payment {
+ type: PaymentType.PIX
 
  pay(amountPayment: number, paymentReference: string ): void {}
